move duplicate mocks at one place
